src/main/java/com/example/Gruppuppgift_Statsbibloteket/controller/LoanController.java
package com.example.Gruppuppgift_Statsbibloteket.controller;

import com.example.Gruppuppgift_Statsbibloteket.Dto.UserLoanDto;
import com.example.Gruppuppgift_Statsbibloteket.model.Loan;
import com.example.Gruppuppgift_Statsbibloteket.service.LoanService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/loans")
public class LoanController {

    @Autowired
    private LoanService loanService;

    // Create loan method
    @PostMapping
    public ResponseEntity<Loan> createLoan(@RequestBody UserLoanDto userLoanDto) {
        try {
            Loan loan = loanService.createLoan(userLoanDto);
            return new ResponseEntity<>(loan, HttpStatus.CREATED);
        } catch (RuntimeException e) {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }

    @GetMapping
    public ResponseEntity<List<Loan>> getAllLoans() {
        List<Loan> loans = loanService.getAllLoans();
        return new ResponseEntity<>(loans, HttpStatus.OK);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Optional<Loan>> deleteLoan(@PathVariable long id) {
        Optional<Loan> loan = loanService.returnBook(id);
        return new ResponseEntity<>(loan, HttpStatus.OK);
    }
}
